treedirstate: ensure appended data is flushed to disk
A dirstate update that appends data to the treedirstate tree file, followed by
a hard reboot before the filesystem cache is flushed, can result in a dirstate
tree root that is referred to by the dirstate file, but does not contain the
correct data.  Ensure the appended data is synced to disk before returning from
`Store.flush()`.

Differential Revision: https://phab.mercurial-scm.org/D1654