repack: exit 0 if a repack is in progress
Previously we raised a LockHeld exception if a repack was already in progress,
which caused hg repack to exit non-zero. This changes it to exit 0 since a
repack is in progress to accomplish the desired result. This is useful in
automation that wants to run repack to maintain the repository but depends on
exit codes to know if a command succeeded or not.

Differential Revision: https://phab.mercurial-scm.org/D1668