mercurial_bundle: bugfix for prematurely closing Compression AsyncRead
Summary:
When using forward/send_all on a Sink it is being flushed and closed, this means that if we call forward multiple times on a Sink it might fail if the underlying IO is already closed.
For ZSTD compression it didn't mattered, because it doesn't use the close/finish, but for other compressions it does.

Reviewed By: jsgf

Differential Revision: D6545343

fbshipit-source-id: 079d950f34716d621dd1fabda4f97d58d0e1d02d