nodemap: implement nodemap in rust
A simple nodemap index (node -> rev, and prefix -> node). See comment in
the code for the actual format.

Note this does not cover every features of a Mecurial nodemap. For example,
it does not have `__setitem__`, `__delitem__`, and does not treat rev -1
specially. Those could be implemented in a higher level.

Differential Revision: https://phab.mercurial-scm.org/D1469