indexes: add comment indication location of python27-sys
The indexes module depends directly on the python27-sys crate for low-level
manipulation of Python structures.  This is located in a subdir in the cpython
git repository.

While that works for git repositories, this causes problems when the crate is
vendored, as Cargo is no longer able to find the directory.

Add a comment to the git URL line with the name of the subdirectory.  This will
be used by the build script to locate the vendored version of the crate.

Differential Revision: https://phab.mercurial-scm.org/D1712