wirepack: return history metadata as a separate stream entry
Summary:
While encoding a wirepack, the entry count for a single filename needs to be
encoded upfront. To avoid buffering in the packer (upcoming diffs), turn this into a stream
entry. The packer will verify that the actual items on the stream match the
entry count provided upfront.

Also remove the path since the caller can now keep track of the current path,
and in that way avoid a path clone for each entry.

Reviewed By: jsgf

Differential Revision: D6577181

fbshipit-source-id: 577bdb9b5fc1a3fadb9b5f7ab4b312afce823ed8