fixup ASAN use-after-free error with fuse request context
Summary:
this is a backport of the same fix from D6578933 and helps
to avoid noise with our new default ASAN enabled build settings.

The root of the problem is that we didn't guarantee to clear the
association between the fuse worker thread and a folly request
associated with a fuse request.  This meant that the lifetime
of the RequestData was being extended until after the EdenMount
was destroyed and then the destructor of RequestData would trigger
an access inside the EdenMount instance.

Adopting the `RequestContextScopeGuard` type helps to avoid this,
at the cost of perhaps prematurely heap allocating a request context
for some opcodes.  D6578933 can avoid that cost.

Reviewed By: chadaustin

Differential Revision: D6602416

fbshipit-source-id: 80785b860af906c0fcf02b42f0235796377c265f