wirepack: support for encoding data entries
Summary:
This is very similar to the support for encoding history entries in the
previous patch. The NULL_HASH wart makes it a little annoying to deal with.

We need to be able to compute how many bytes a delta has efficiently -- add a
function and tests for that as well.

Reviewed By: jsgf

Differential Revision: D6589117

fbshipit-source-id: e3a05672e1e8f455a2e95b30418468d1e3254407