infinitepush: fix infinitepush pull for cgversion 3
Summary:
Previously if only cgversion 3 is supported by the server, then hg up HASH
may fail if _rebundle() is needed.
This diff fixes it by implementing the same logic as upstream mercurial. It
uses bundle capabilities to find out which versions client supports and it
intersects it with versions that are supported by the server.

This is grafted from D6655986

Test Plan:
Tested in fb-hgext: ran unit-test

Reviewers:

Subscribers:

Tasks:

Tags: