infinitepush: move sql related infinitepush tests from fb-hgext/tests to tests
Summary:
Now that we moved infinitepush to hgext in D6691670, we can also move
the tests. The sql related tests need special attention because of the db
connection and mysql connector.

Test Plan:
Ran all the tests. Removed the mysql connector and checked that the
test is skipped. Also checked that the test is skipped if the db connection
details are not specified.

Reviewers: ikostia, singhsrb, #mercurial, #sourcecontrol

Reviewed By: ikostia

Subscribers: rmcelroy

Differential Revision: https://phabricator.intern.facebook.com/D6691821

Signature: 6691821:1515600428:279dc4796c9be721c5fa81ac2dcd6e065ac30f18