phabricator-graphql: nice error message if ~/.arcrc is bad
Summary: When running hg ssl, if the .arcrc file is not proper JSON
(e.g., has a trailing comma), json.loads raises an exception that
was not being caught and lead to a stacktrace.

Test Plan: Locally run hg ssl with a bag .arcrc and the stack trace was gone.
Also write a test in test-fb-hgext-phabstatus.t

Reviewers: simonfar

Subscribers: luizmramos

Tasks: T24993088

Tags: