pushrebase: correctly track lastdestnode when rebasing merges
Summary:
The initial value of lastdestnode should be the commit onto which the rebase is
happening.

It should be updated whenever a new commit is added for which the p1 parent is
the old last dest node.

Test Plan: Run the new test, and show it now works.

Reviewers: durham, rmcelroy, #mercurial

Reviewed By: rmcelroy

Subscribers: mitrandir

Differential Revision: https://phabricator.intern.facebook.com/D6709108

Signature: 6709108:1515760332:32bef9892792c7c94117b8fe14364937d98ebbac