pushrebase: add test showing failure to handle p2-parent merges
Summary:
If pushrebase is asked to rebase a commit where the p2-parent is the route to
the commit onto which the rebase is being made, then it is supposed to flip the
parents.  This is what ensures that the changes from the server-only commits
are not lost.

This doesn't work.  In this test, the "other" file, created on the server, is
lost in the final merged commit.

Test Plan: Added a test which demonstrates a failure.

Reviewers: quark, #mercurial

Reviewed By: quark

Subscribers: quark

Differential Revision: https://phabricator.intern.facebook.com/D6709096

Signature: 6709096:1515794007:0908765c24b18a7dcaf3672953e816095b0acf68