daemon should chdir / to avoid circular issues
Summary:
akushner mentioned that he saw a hang when performing
a graceful restart and we realized that his cwd was inside the
mount point that he was taking over.  While this doesn't guarantee
problems (I've run it this way almost exclusively with no issues)
it feels like a potential problem to hold a reference to a file
descriptor within a fuse mount that we are ourselves going to
manage: it seems likely that we could end up deadlocking if
we were to attempt any IO on such a descriptor.

This diff seeks to head off this particular case by simply
chdiring to `/` very early on startup.  This is typically
something that is handled by the `daemonize` function in
old school unix code.

Reviewed By: simpkins

Differential Revision: D6740547

fbshipit-source-id: 986004e9ed6cfa493ad55c5cdbca1805fc0f7768