better fix for type error
Summary:
On closer investigation, I realized that the typechecker is
complaining about a real problem here, that was apparently hidden by
the stubs. The problem is explained in
https://our.intern.facebook.com/intern/dex/qa/1596320410425583/

It's not safe to `import thrift` followed by accessing
`thrift.Thrift.TException`; it only happens to work at runtime if some
intervening code has explicitly imported `thrift.Thrift` submodule for
you.

This diff fixes the code to be safer and also make mypy happy.

Reviewed By: chadaustin

Differential Revision: D6761089

fbshipit-source-id: 5c6fdfa0bef893c4aea17ade8e7bfacaf7c0c8cc