mononoke: fix lookup command param parsing
Summary:
Previously parsing code assumed that lookup param is an ident string. This is
incorrect, because ident string accepts only strings that consist only of
digits and small letters and start with a letter. Since a hash can be a lookup
param, this param may start with non-letter. More importantly, any bookmark can
be a lookup parameter. That means that this param can have spaces and
punctuation marks.

Let's use simpler parsing that just parses the whole string.

Reviewed By: jsgf

Differential Revision: D6758916

fbshipit-source-id: de2dd144cebfe88c751e8110ef6e3a71d3769800