clindex: fix issues on small repos and turn on by default
Summary:
This fixes an obvious mistake. It wasn't noticed because it's part of the last
"cleanup" before landing the change. And it runs on a big repo which hides
the problem.

On smaller repos, changelog is inlined and clindex won't be used because the
Rust index does not support inlined revlogs.  `repo.updatecaches()` will raise
`AttributeError` since the original `self.changelog.index` object does not have
`updatecaches()`. In that case, it's obvious that we should not skip the
original `repo.updatecaches()` call.  The same applies to `repo.destorying()`.

Since after the fix, tests are passing. Let's enable clindex by default. As
explained above, clindex is not actually well exercised because of inlined
revlog.  We should probably special case changelog to be always non-inlined.
That's probably a separate diff.

Reviewed By: phillco

Differential Revision: D6767200

fbshipit-source-id: c8432eec6c02c06c4d49c13df150e1b25bbaa1aa